move AI endpoints to its own file
